fix: simplify typing for user events

+++ app/api/clerk/deleteUser/route.tsx
@@ -1,22 +1,21 @@
-import { DeletedObjectJSON, UserWebhookEvent } from '@clerk/nextjs/server'
-
 import { deleteUser } from '@/app/app/actions'
 import { validateRequest } from '@/lib/webhookUtils'
 
 const webhookSecret = process.env.CLERK_WEBHOOK_SECRET_UPDATE_USER || ``
 
 export async function POST(request: Request) {
-  const payload = (await validateRequest(
-    request,
-    webhookSecret,
-  )) as UserWebhookEvent
+  const payload = await validateRequest(request, webhookSecret)
   console.log(payload)
-  const data = payload.data as DeletedObjectJSON
 
-  const id = data.id
+  if (payload.type !== 'user.deleted') {
+    return new Response('Not a user deleted event', { status: 400 })
+  }
+
+  const id = payload.data.id
   if (!id) {
     return new Response('Missing user ID', { status: 400 })
   }
+
   const user = await deleteUser(id)
   console.log({ user })
   if (!user) {
